Add support for multiple vCPUs

+++ src/arch/aarch64/smc.c
@@ -192,7 +192,7 @@ bool smc_register_sip_handler(smc_sip_handler_t handler)
 }
 
 // @ivanv: print out which SMC call as a string we can't handle.
-bool smc_handle(size_t vcpu_id, uint32_t hsr)
+bool smc_handle(size_t vcpu_id, uint64_t hsr)
 {
     seL4_UserContext regs;
     int err = seL4_TCB_ReadRegisters(BASE_VM_TCB_CAP + vcpu_id, false, 0, SEL4_USER_CONTEXT_SIZE, &regs);
